bibl/image_plus_text/pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"

[project]
name = "imageplustxt"
authors = [{name = "surumkata", email = "a93277@uminho.pt"}]
license = {file = "LICENSE"}
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = ["Pillow","argparse"]
readme = "readme.md"

[project.scripts]
ipt = "imageplustxt:imageplustxt"